fix(Transfer): 修复双向穿梭时，checkbox 数据更新异常
