+++ components/transfer/style/index.less
@@ -51,9 +51,9 @@
         }
 
         .@{transfer}-panel-list {
-            // Tree 本身自带一点左边距，Transfer 反向缩回去
-            margin-left: calc(0px - @option-inline-padding);
             width: calc(100% + @option-inline-padding * 2);
+            // Tree 本身自带一点左边距，Transfer 反向缩回去
+            margin-left: calc(0px - @option-inline-padding); 
 
             .@{transfer}-option {
                 height: @option-height;
