+++ components/transfer/useOptionsFilter.ts
@@ -1,0 +1,28 @@
+import { type Ref, computed, ref } from 'vue';
+import {
+    type TransferFilter,
+    type TransferInjection,
+    type TransferOption,
+} from './interface';
+
+export const useOptionsFilter = ({
+    options,
+    filter,
+    rootProps,
+}: Pick<TransferInjection, 'rootProps'> & {
+    options: Ref<TransferOption[]>;
+    filter: TransferFilter;
+}) => {
+    const filterText = ref<string>('');
+
+    const displayOptions = computed<TransferOption[]>(() => {
+        if (!rootProps.filterable) {
+            return options.value;
+        }
+        return options.value.filter((option) =>
+            filter(filterText.value, option),
+        );
+    });
+
+    return { filterText, displayOptions };
+};
